attempt to "group" network transmissions
