+++ Source/ACE.Server/Network/NetworkSession.cs
@@ -25,8 +25,10 @@ namespace ACE.Server.Network
 
         private readonly Session session;
 
-        private readonly Object currentBundleLock = new Object();
-        private NetworkBundle currentBundle = new NetworkBundle();
+        //private readonly Object currentBundleLock = new Object();
+        private ConcurrentDictionary<GameMessageGroup, Object> currentBundleLocks = new ConcurrentDictionary<GameMessageGroup, Object>();
+        private ConcurrentDictionary<GameMessageGroup, NetworkBundle> currentBundles = new ConcurrentDictionary<GameMessageGroup, NetworkBundle>();
+        //private NetworkBundle currentBundle = new NetworkBundle();
 
         private ConcurrentDictionary<uint, ClientPacket> outOfOrderPackets = new ConcurrentDictionary<uint, ClientPacket>();
         private ConcurrentDictionary<uint, MessageBuffer> partialFragments = new ConcurrentDictionary<uint, MessageBuffer>();
@@ -78,6 +80,13 @@ namespace ACE.Server.Network
             ServerId = serverId;
             // New network auth session timeouts will always be low.
             TimeoutTick = DateTime.UtcNow.AddSeconds(AuthenticationHandler.DefaultAuthTimeout).Ticks;
+
+            foreach (var gmg in System.Enum.GetValues(typeof(GameMessageGroup)))
+            {
+                var group = (GameMessageGroup)gmg;
+                currentBundleLocks[group] = new object();
+                currentBundles[group] = new NetworkBundle();
+            }
         }
 
         /// <summary>
@@ -87,17 +96,24 @@ namespace ACE.Server.Network
         /// <param name="messages">One or more GameMessages to send</param>
         public void EnqueueSend(params GameMessage[] messages)
         {
-            lock (currentBundleLock)
+            messages.GroupBy(k => k.Group).ToList().ForEach(k =>
             {
-                currentBundle.EncryptedChecksum = true;
-
-                foreach (var message in messages)
+                var grp = k.First().Group;
+                var currentBundleLock = currentBundleLocks[grp];
+                lock (currentBundleLock)
                 {
-                    log.DebugFormat("[{0}] Enqueuing Message {1}", session.LoggingIdentifier, message.Opcode);
-                    currentBundle.Enqueue(message);
+                    foreach (var msg in k)
+                    {
+                        var currentBundle = currentBundles[msg.Group];
+                        currentBundle.EncryptedChecksum = true;
+                        log.DebugFormat("[{0}] Enqueuing Message {1}", session.LoggingIdentifier, msg.Opcode);
+                        currentBundle.Enqueue(msg);
+                    }
                 }
-            }
+            });
         }
+
+
 
         /// <summary>
         /// Enqueues a ServerPacket for sending to this client.
@@ -121,42 +137,61 @@ namespace ACE.Server.Network
         public void Update(double lastTick)
         {
             ConnectionData.ServerTime += lastTick;
-            NetworkBundle bundleToSend = null;
-            lock (currentBundleLock)
+
+            currentBundles.Keys.ToList().ForEach(group =>
             {
-                if (sendResync && !currentBundle.TimeSync && DateTime.UtcNow > nextResync)
+                var currentBundleLock = currentBundleLocks[group];
+                var currentBundle = currentBundles[group];
+
+                NetworkBundle bundleToSend = null;
+                lock (currentBundleLock)
                 {
-                    log.DebugFormat("[{0}] Setting to send TimeSync packet", session.LoggingIdentifier);
-                    currentBundle.TimeSync = true;
-                    currentBundle.EncryptedChecksum = true;
-                    nextResync = DateTime.UtcNow.AddMilliseconds(timeBetweenTimeSync);
+                    if (group == GameMessageGroup.InvalidQueue)
+                    {
+                        if (sendResync && !currentBundle.TimeSync && DateTime.UtcNow > nextResync)
+                        {
+                            log.DebugFormat("[{0}] Setting to send TimeSync packet", session.LoggingIdentifier);
+                            currentBundle.TimeSync = true;
+                            currentBundle.EncryptedChecksum = true;
+                            nextResync = DateTime.UtcNow.AddMilliseconds(timeBetweenTimeSync);
+                        }
+
+                        if (sendAck && !currentBundle.SendAck && DateTime.UtcNow > nextAck)
+                        {
+                            log.DebugFormat("[{0}] Setting to send ACK packet", session.LoggingIdentifier);
+                            currentBundle.SendAck = true;
+                            nextAck = DateTime.UtcNow.AddMilliseconds(timeBetweenAck);
+                        }
+
+                        if (currentBundle.NeedsSending && DateTime.UtcNow > nextSend)
+                        {
+                            log.DebugFormat("[{0}] Swaping bundle", session.LoggingIdentifier);
+                            // Swap out bundle so we can process it
+                            bundleToSend = currentBundle;
+                            currentBundles[group] = new NetworkBundle();
+                        }
+                    }
+                    else
+                    {
+                        if (currentBundle.NeedsSending && DateTime.UtcNow > nextSend)
+                        {
+                            log.DebugFormat("[{0}] Swaping bundle", session.LoggingIdentifier);
+                            // Swap out bundle so we can process it
+                            bundleToSend = currentBundle;
+                            currentBundles[group] = new NetworkBundle();
+                        }
+                    }
                 }
 
-                if (sendAck && !currentBundle.SendAck && DateTime.UtcNow > nextAck)
+                // Send our bundle if we have one
+                // We should be able to execute this outside the lock as Sending is single threaded
+                // and all future writes from other threads will go to the new bundle
+                if (bundleToSend != null)
                 {
-                    log.DebugFormat("[{0}] Setting to send ACK packet", session.LoggingIdentifier);
-                    currentBundle.SendAck = true;
-                    nextAck = DateTime.UtcNow.AddMilliseconds(timeBetweenAck);
+                    SendBundle(bundleToSend, group);
+                    nextSend = DateTime.UtcNow.AddMilliseconds(minimumTimeBetweenBundles);
                 }
-
-                if (currentBundle.NeedsSending && DateTime.UtcNow > nextSend)
-                {
-                    log.DebugFormat("[{0}] Swaping bundle", session.LoggingIdentifier);
-                    // Swap out bundle so we can process it
-                    bundleToSend = currentBundle;
-                    currentBundle = new NetworkBundle();
-                }
-            }
-
-            // Send our bundle if we have one
-            // We should be able to execute this outside the lock as Sending is single threaded
-            // and all future writes from other threads will go to the new bundle
-            if (bundleToSend != null)
-            {
-                SendBundle(bundleToSend);
-                nextSend = DateTime.UtcNow.AddMilliseconds(minimumTimeBetweenBundles);
-            }
-
+            });
             FlushPackets();
         }
 
@@ -419,8 +454,10 @@ namespace ACE.Server.Network
             }
         }
 
+        //is this special channel
         private void FlagEcho(float clientTime)
         {
+            var currentBundle = currentBundles[GameMessageGroup.InvalidQueue];
             lock (currentBundle)
             {
                 // Debug.Assert(clientTime == -1f, "Multiple EchoRequests before Flush, potential issue with network logic!");
@@ -475,7 +512,7 @@ namespace ACE.Server.Network
                 else
                     packet.Header.Sequence = ConnectionData.PacketSequence.NextValue;
                 packet.Header.Id = ServerId;
-                packet.Header.Table = 0x14;
+                //packet.Header.Table = 0x14; //comment this for the client confusion, uncomment it and the "GameMessageGroup" work becomes useless.
                 packet.Header.Time = (ushort)ConnectionData.ServerTime;
 
 
@@ -524,7 +561,7 @@ namespace ACE.Server.Network
         /// several packets as needed.
         /// </summary>
         /// <param name="bundle"></param>
-        private void SendBundle(NetworkBundle bundle)
+        private void SendBundle(NetworkBundle bundle, GameMessageGroup group)
         {
             log.DebugFormat("[{0}] Sending Bundle", session.LoggingIdentifier);
             bool writeOptionalHeaders = true;
@@ -535,6 +572,7 @@ namespace ACE.Server.Network
             while (bundle.HasMoreMessages)
             {
                 var message = bundle.Dequeue();
+
                 var fragment = new MessageFragment(message, ConnectionData.FragmentSequence++);
                 fragments.Add(fragment);
             }
@@ -620,7 +658,7 @@ namespace ACE.Server.Network
                         availableSpace -= (uint)packet.Data.Length;
                     }
                 }
-
+                packet.Header.Table = (ushort)group;
                 EnqueueSend(packet);
             }
         }
@@ -686,98 +724,6 @@ namespace ACE.Server.Network
                 }
                 stream.Seek(0, SeekOrigin.Begin);
                 return new ClientMessage(stream);
-            }
-        }
-
-        private class MessageFragment
-        {
-            public GameMessage Message { get; private set; }
-
-            public uint Sequence { get; set; }
-
-            public ushort Index { get; set; }
-
-            public ushort Count { get; set; }
-
-            public uint DataLength => (uint)Message.Data.Length;
-
-            public uint DataRemaining { get; private set; }
-
-            public uint NextSize
-            {
-                get
-                {
-                    var dataSize = DataRemaining;
-                    if (dataSize > ServerPacketFragment.MaxFragmentDataSize)
-                        dataSize = ServerPacketFragment.MaxFragmentDataSize;
-                    return PacketFragmentHeader.HeaderSize + dataSize;
-                }
-            }
-
-            public uint TailSize => PacketFragmentHeader.HeaderSize + (DataLength % ServerPacketFragment.MaxFragmentDataSize);
-
-            public bool TailSent { get; private set; }
-
-            public MessageFragment(GameMessage message, uint sequence)
-            {
-                Message = message;
-                DataRemaining = DataLength;
-                Sequence = sequence;
-                Count = (ushort)(Math.Ceiling((double)DataLength / PacketFragment.MaxFragmentDataSize));
-                Index = 0;
-                if (Count == 1)
-                    TailSent = true;
-                log.DebugFormat("Sequence {0}, count {1}, DataRemaining {2}", sequence, Count, DataRemaining);
-            }
-
-            public ServerPacketFragment GetTailFragment()
-            {
-                var index = (ushort)(Count - 1);
-                TailSent = true;
-                return CreateServerFragment(index);
-            }
-
-            public ServerPacketFragment GetNextFragment()
-            {
-                return CreateServerFragment(Index++);
-            }
-
-            private ServerPacketFragment CreateServerFragment(ushort index)
-            {
-                log.DebugFormat("Creating ServerFragment for index {0}", index);
-                if (index >= Count)
-                    throw new ArgumentOutOfRangeException("index", index, "Passed index is greater then computed count");
-
-                var position = index * ServerPacketFragment.MaxFragmentDataSize;
-                if (position > DataLength)
-                    throw new ArgumentOutOfRangeException("index", index, "Passed index computes to invalid position size");
-
-                if (DataRemaining <= 0)
-                    throw new InvalidOperationException("There is no data remaining");
-
-                var dataToSend = DataLength - position;
-                if (dataToSend > ServerPacketFragment.MaxFragmentDataSize)
-                    dataToSend = ServerPacketFragment.MaxFragmentDataSize;
-
-                if (DataRemaining < dataToSend)
-                    throw new InvalidOperationException("More data to send then data remaining!");
-
-                // Read data starting at position reading dataToSend bytes
-                Message.Data.Seek(position, SeekOrigin.Begin);
-                byte[] data = new byte[dataToSend];
-                Message.Data.Read(data, 0, (int)dataToSend);
-
-                // Build ServerPacketFragment structure
-                ServerPacketFragment fragment = new ServerPacketFragment(data);
-                fragment.Header.Sequence = Sequence;
-                fragment.Header.Id = 0x80000000;
-                fragment.Header.Count = Count;
-                fragment.Header.Index = index;
-                fragment.Header.Group = (ushort)Message.Group;
-
-                DataRemaining -= dataToSend;
-                log.DebugFormat("Done creating ServerFragment for index {0}. After reading {1} DataRemaining {2}", index, dataToSend, DataRemaining);
-                return fragment;
             }
         }
